backoffice: loan tests
